Initial commit
Initial commit

Webpage is in a state in which it is more or less ready to be made
public. Things that don't work yet (among others) are:

- views in pages apart from achievements
- tags (and their links)
- search
- annual reports and publication lists only selectors in achievements
- Individual member pages

Pages still missing are:

- Annual reports and publication lists from 2011 onward
- Cleaning up of achievements posts
- Posts on new papers (scripts need to be made more versatile to be able
to handle individual bibtex entries)

Trawling through the old webpage directory tree, there are lots of pages
not directly mapped onto the new site. Things that may be worth
including are more previous conferences, like the Hayashi lecture, the
Rironcon symposium 2012.

Need to decide:

- Whether to use Travis
- Whether to use Prose.io
- .. or just use gorilla as a staging server.