Tentaikeisei program generated by TAG_conferences python script
When writing times in ruamel.yaml, no quotes are added around the time. It's impossible to add quotes around them, they keep getting dismissed, even with preserve_quotes option. So we add a "!" at the beginning, then quotes are included (because "!" is a special symbol in yaml, a yaml tag. Then we remove the "!" in the jekyll layout. It's the only workaround i found after trying for HOURS!