#back-top hack
back-top doesn't work on the CCS server, but it works on my machine locally. The back-top button doesn't seem to fade in properly, but i've spent hours unsuccessfully searching for the reason, so we're just not hiding it initially, and it'll always be there, even if we're at the top of the page. The back-top is very important, for example, for conference pages to get back to the table of contents.