Remove `\\` from second line in latex data
This is because individual members are `\input{}` into `meibo.tex` followed by a `\midrule`, and `\midrule` must always be preceded by `\\`. Strangely, the LuaLaTeX complier throws an error if the preceding `\\` is in the input file, but it is alright if it is in `meibo.tex` just before  \midrule. It is not clear why, but it works this way, so we make this change.